+++ helion/autotuner/differential_evolution.py
@@ -150,7 +150,7 @@ class DifferentialEvolutionSearch(PopulationBasedSearch):
         # The initial population is 2x larger so we can throw out the slowest half and give the tuning process a head start
         initial_population_name = self.initial_population_strategy.name
         oversized_population = sorted(
-            self.parallel_benchmark_flat(
+            self.benchmark_flat_batch(
                 self._generate_initial_population_flat(),
             ),
             key=performance,
@@ -167,7 +167,7 @@ class DifferentialEvolutionSearch(PopulationBasedSearch):
         if not indices:
             return []
         flat_configs = [self.mutate(i) for i in indices]
-        return self.parallel_benchmark_flat(flat_configs)
+        return self.benchmark_flat_batch(flat_configs)
 
     def iter_candidates(self) -> Iterator[tuple[int, PopulationMember]]:
         if self.immediate_update:
